+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased (0.13.0)]
+### Added
+- A way to obtain where tilesets, templates and maps have been loaded from
+
 ## [0.12.0]
 ### Added
 - Add `text`, `width` and `height` members to `ObjectShape::Text`. (#278)