+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased (0.12.2)]
+### Fixed
+- Fixed template instance size and position overrides in `ObjectData::shape`. (#309)
+
 ## [0.12.1]
 ### Changed
 - Improved documentation on `Map::layers` and `Map::get_layer`. (#306)